+++ sync-for-payroll/src/sdk/sdk.ts
@@ -60,9 +60,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "1.3.0";
-    genVersion = "2.139.1";
-    userAgent = "speakeasy-sdk/typescript 1.3.0 2.139.1 3.0.0 @codat/sync-for-payroll";
+    sdkVersion = "1.4.0";
+    genVersion = "2.143.2";
+    userAgent = "speakeasy-sdk/typescript 1.4.0 2.143.2 3.0.0 @codat/sync-for-payroll";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);